requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.1.0
blinker==1.7.0
certifi==2023.11.17
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==2.3.3
frozenlist==1.4.1
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.27.9
packaging==23.2
pdf2image==1.16.3
Pillow==10.0.0
PyMuPDF==1.23.3
PyMuPDFb==1.23.3
pytesseract==0.3.10
python-dotenv==1.0.0
requests==2.31.0
rfc3986==1.5.0
sniffio==1.3.0
tqdm==4.66.1
urllib3==2.1.0
waitress==2.1.2
Werkzeug==3.0.1
yarl==1.9.4
